requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
black==24.4.2
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.3
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
GDAL==3.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
morecantile==5.3.0
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pyproj==3.6.1
pystac==1.10.1
python-dateutil==2.9.0.post0
rasterio==1.3.10
rio-tiler==6.6.1
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
tifffile==2024.5.22
tiffile==2018.10.18
typing_extensions==4.11.0
wheel==0.43.0
